Initialize DeltaSharingRestClient in its own file
